REFACTOR (Controllers): Use App\Http\Controllers\Controller; and update namespaces.
